feat-basket: covert price from basketcheckout v1 v2 to decimal

+++ Infrastructure/EventBus.Messages/Events/BasketCheckoutEventV2.cs
@@ -3,7 +3,7 @@
     public class BasketCheckoutEventV2 : BaseIntegrationEvent
     {
         public string? UserName { get; set; }
-        public double? TotalPrice { get; set; }
+        public decimal? TotalPrice { get; set; }
 
     }
 }
